Restore the 2.3 editions of the Notebooks so that they will appear on Colab. (#26)
Older versions of the documentation have links to these files
at their hosted locations on Colab. This commit restores those
files to Colab.

In the future, we'll use version-specific links to the Colab
notebooks. For 2.3 and earlier, readers will get the latest
version available on Colab (usually 2.3).

All files are the 2.3 version, except for these two which
are 2.2 editions (no newer version exists under these names):

- RIME_Bias_and_Fairness_Walkthrough_Lending.ipynb (in 2.3 it became RIME_Lending_Classification_Walkthrough.ipynb )
- RIME_Firewall_Configuring.ipynb (in 2.3 it became RIME_Continuous_Test_Configuring.ipynb )